src/ai/requirements.txt
python-dotenv
openai
gradio
faiss-cpu
#faiss-gpu
langchain-faiss
langchain-openai
langchain-huggingface
nltk
rank_bm25